+++ Core/Libraries/Source/WWVegas/WWMath/matrix3d.cpp
@@ -361,8 +361,8 @@ void Matrix3D::Look_At(const Vector3 &p,const Vector3 &t,float roll)
 	dy = (t[1] - p[1]);
 	dz = (t[2] - p[2]);
 
-	len1 = (float)WWMath::Sqrt(dx*dx + dy*dy + dz*dz);
-	len2 = (float)WWMath::Sqrt(dx*dx + dy*dy);
+	len1 = WWMath::Sqrt(dx*dx + dy*dy + dz*dz);
+	len2 = WWMath::Sqrt(dx*dx + dy*dy);
 
 	if (len1 != 0.0f) {
 		sinp = dz/len1;
