+++ GeneralsMD/Code/Libraries/Source/WWVegas/WW3D2/meshgeometry.cpp
@@ -1395,7 +1395,7 @@ void MeshGeometryClass::Compute_Bounds(Vector3 * verts)
 
 	// calculate the bounding sphere
 	BoundSphereCenter = (BoundBoxMin + BoundBoxMax)/2.0f;
-	BoundSphereRadius = (float)(BoundBoxMax-BoundSphereCenter).Length2();
+	BoundSphereRadius = (BoundBoxMax-BoundSphereCenter).Length2();
 	BoundSphereRadius = ((float)sqrt(BoundSphereRadius))*1.00001f;
 	Set_Flag(DIRTY_BOUNDS,false);
 }
